change to 64 to match duration. error handle

+++ turbo-genesis-ffi/src/audio/native.rs
@@ -20,4 +20,4 @@ pub fn get_volume(key_ptr: *const u8, key_len: u32) -> f32 {
 
 pub fn set_volume(key_ptr: *const u8, key_len: u32, decibels: f32) {}
 
-pub fn seek_to(key_ptr: *const u8, key_len: u32, seconds: f32) {}
+pub fn seek_to(key_ptr: *const u8, key_len: u32, seconds: f64) {}
